The Matrix
The Matrix shows the top 30 product brands with their total transactions, total profits, profit margin, and the return rate of the products between the years 1997 and 1998. The Total transactions are colour coded from the most transacted to the least transacted products. This is shown by the data bars.

Additionally, the Profit margin and the return rates are colour coded as well, with the background colors: White to green for the profit margin, with green being the highest where whereas the return rate is white to red, with red being the highest rate. The lower the return rate the better.