The Final Dashboard
This showcases the final output of my visualization. The matrix was generated with my top 30 products sold, the map showing the 3 major markets that the business is focused on, which can be drilled down to various states and cities for further analysis. 

Additionally, it captures the very essential KPIs, the tree map, the charts, the gauge, and the filters applied to them.

Finally, some bookmarks are also included to capture the required information.