The KPIs
The very essential KPIs: 

This includes the current month transactions, the current month profits, and the current month returns. The KPIs are visualized against their previous month's targets, with the smaller graphs within them showing the trend taken from the previous month. They also capture the deviation, either positive or negative growth from the previous month s  values. This shows a clear trend in the business direction.

Note that this are arrived at sing the measures created and stored in the measures table of my dataset.